Extract `LoaderSpy` class in its own file
